refactor: move Accounts code to accounts plugin
